Upload some fixes in remove cantidad from inventario_table

+++ database/migrations/2025_11_09_214014_remove_cantidad_from_inventario_biblioteca_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('inventario_biblioteca', function (Blueprint $table) {
-            $table->dropColumn('cantidad');
+            if (Schema::hasColumn('inventario_biblioteca', 'cantidad')) {
+                $table->dropColumn('cantidad');
+            }
         });
     }
